requirements.txt
annotated-types==0.6.0
aws-lambda-powertools==2.38.1
dateutils==0.6.12
dnspython==2.6.1
environ==1.0
mongoengine==0.20.0
mongoengine_goodjson==1.1.8
pydantic==2.7.1
pydantic_core==2.18.2
pymongo==3.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
typing_extensions==4.11.0
